Add abstract class TimeSeriesChart
