+++ src/d3/xy-chart.d3.ts
@@ -2,16 +2,16 @@ import * as d3 from 'd3';
 import { BaseD3, RenderOptions as BaseRenderOptions } from './base.d3';
 import { Observable } from 'rxjs';
 import { formatX } from '../utils/formatters';
-import { TimeSeriesPoint } from '../datasets/metas/types';
+import { XYDatum } from '../datasets/metas/types';
 import { TimeSeriesDatum } from '../datasets/queries/time-series.query';
 
-export interface RenderOptions<Datum> extends BaseRenderOptions {
+export interface RenderOptions<Point, Datum extends XYDatum<Point>> extends BaseRenderOptions {
   data$: Observable<Datum[]>;
-  activePoint$: Observable<TimeSeriesPoint | null>;
+  activePoint$: Observable<Point | null>;
 }
 
-export abstract class XYChartD3<LegendItemStyle,
-  Datum extends TimeSeriesDatum<LegendItemStyle> = TimeSeriesDatum<LegendItemStyle>> extends BaseD3<RenderOptions<Datum>> {
+export abstract class XYChartD3<Point, Datum extends XYDatum<Point>, ScaleX = d3.ScaleLinear<number, number> | d3.ScaleBand<string>
+  > extends BaseD3<RenderOptions<Point, Datum>> {
   static padding = 16;
   static yAxisWidth = 40;
   static xAxisHeight = 20;
@@ -22,7 +22,7 @@ export abstract class XYChartD3<LegendItemStyle,
   static fontSizeSmall = 12;
   static fontSizeMedium = 14;
 
-  protected scaleX: d3.ScaleTime<number, number>;
+  protected scaleX: ScaleX;
   protected scaleY: d3.ScaleLinear<number, number>;
   protected xAxis: d3.Axis<Date>;
   protected yAxis: d3.Axis<number>;
@@ -30,14 +30,6 @@ export abstract class XYChartD3<LegendItemStyle,
   protected yAxisG: d3.Selection<SVGGElement, unknown, null, undefined>;
   protected dataG: d3.Selection<SVGGElement, unknown, null, undefined>;
   protected legendG: d3.Selection<SVGGElement, unknown, null, undefined>;
-
-  x(value: Date) {
-    return this.scaleX(value);
-  }
-
-  y(value: number) {
-    return this.scaleY(value);
-  }
 
   render() {
     super.render();
@@ -77,6 +69,13 @@ export abstract class XYChartD3<LegendItemStyle,
       .attr('transform', `translate(${left},${top})`);
   }
 
+
+  protected renderData() {
+    this.dataG = this.svg
+      .append('g')
+      .attr('class', 'xy_chart-data');
+  }
+
   protected updateLegend(data: Datum[]) {
     const { legendIconWidth, fontSizeMedium } = XYChartD3;
 
@@ -100,67 +99,15 @@ export abstract class XYChartD3<LegendItemStyle,
     }
   }
 
-  protected renderAxis() {
-    const { height, width } = this.renderOptions;
-    const { xAxisHeight, yAxisWidth, legendHeight, padding, fontSizeSmall } = XYChartD3;
+  protected abstract renderAxis();
 
-    const top = padding;
-    const bottom = height - xAxisHeight - legendHeight - padding;
-    const left = padding + yAxisWidth;
-    const right = width - padding;
-
-    this.scaleX = d3
-      .scaleUtc()
-      .range([left, right]);
-    this.scaleY = d3
-      .scaleLinear()
-      .nice()
-      .range([bottom, top]);
-
-    this.xAxis = d3
-      .axisBottom<Date>(this.scaleX)
-      .ticks(d3.timeWeek.every(1))
-      .tickFormat(formatX);
-    this.yAxis = d3.axisLeft<number>(this.scaleY);
-
-    this.xAxisG = this.svg
-      .append('g')
-      .attr('class', 'xy_chart-x_axis')
-      .attr('transform', `translate(0,${bottom})`)
-      .attr('font-size', fontSizeSmall);
-    this.yAxisG = this.svg
-      .append('g')
-      .attr('class', 'xy_chart-y_axis')
-      .attr('transform', `translate(${left},0)`)
-      .attr('font-size', fontSizeSmall);
-  }
-
-  protected updateAxis(data: Datum[]) {
-    const points = data.flatMap(datum => datum.points);
-    const xs = points.map(point => point.x);
-    const ys = points.map(point => point.y);
-    this.scaleX.domain([d3.min(xs)!, d3.max(xs)!]);
-    this.scaleY.domain([0, d3.max(ys)!]);
-
-    this.xAxisG
-      .transition(this.transition)
-      .call(this.xAxis);
-    this.yAxisG
-      .transition(this.transition)
-      .call(this.yAxis);
-  }
-
-  protected renderData() {
-    this.dataG = this.svg
-      .append('g')
-      .attr('class', 'xy_chart-data');
-  }
+  protected abstract updateAxis(data: Datum[]);
 
   protected abstract updateData(data: Datum[]);
 
   protected abstract renderActivePoint();
 
-  protected abstract updateActivePoint(activePoint: TimeSeriesPoint | null);
+  protected abstract updateActivePoint(activePoint: Point | null);
 
   protected abstract appendLegendItemIcon(itemG: d3.Selection<SVGGElement, unknown, null, undefined>, datum: Datum);
 }
