+++ src/datasets/metas/types.ts
@@ -3,6 +3,7 @@ import { TabbedChartsMeta } from './tabbed-charts.meta';
 import { VRScatterplotMeta } from './vr-scatter-plot.meta';
 import { GeoMapMeta } from './geo-map.meta';
 import { A11yModuleImporter } from '../../directives/a11y/a11y.directive';
+import { SummaryGroup } from '../summarizations/types';
 
 export interface BaseMeta<T extends MetaType> {
   type: T;
@@ -29,6 +30,13 @@ export interface ScatterPlotMeta<T extends MetaType, QueryT> extends DataMeta<T,
   yLabel?: string;
   zLabel?: string;
 }
+
+export interface XYDatum<Point> {
+  label: string;
+  points: Point[];
+  querySummaries?: () => SummaryGroup[];
+}
+
 export type TimeSeriesPoint = XYPoint<Date, number>;
 
 export type NumPoint = XYPoint<number, number>;
